Provision by re-using terminated instances
When there's a need for an extra instance, first check
if there's a terminated instance in GCE that is not associated
with any Jenkins node. If so, start that instance, and associate
the newly-created Jenkins node with that instance.

This works in simple cases. It ignores instances that are
currently stopping. This is probably good, but, the plugin
probably doesn't re-provision at the moment when an instance
completes the stop operation. The plugin also probably doesn't
limit the max number of persistent instances in a group -
this would probably be useful.