Persist a configurable number of instances per group
- Allow user to choose number of instances to persist
- When terminating nodes, persist (stop) or delete instance
  depending on current instance count

This strategy works OK except that if several nodes in
the same config are being deleted at the same time,
and there are more instances than are allowed to be
persisted, no instances are persisted until the number
of nodes (including nodes that are stopping) are
below the max-instances-to-persist count. This can
result in intances getting deleted when they should
get persisted.

This is not a problem for "max 2, persist 2" type
configurations, but causes "max 10, persist 2" type
configurations to soemtimes persist less than 2.