Instance termination is a two-phase process
First stop instance. Wait for it to complete.
Then delete it (if it should not be persisted). Wait for it to complete.

This moves the decision "persist or not?" to the right time
-- multiple overlapping instance decommissions will not
lead to deleting more instances than necessary.

All the async work happens on the remoting thread pool.

The async logic is not robust against the clouds/instance configs/etc
changing in parallel.

The async logic has rather poor error handling as well.