Enqueue items in operation tracker before issuing API calls
Sometimes, when multiple instances are decommissioned at
the same time, multiple threads will find out that the instance
each is waiting on has stopped at nearly the same time
(less than a second apart). When this happens,
all these will delete instances, because they are not aware
that the other threads are also issuing delete operations.

To avoid this, we will put deletes into the operation tracker
before performing API calls. THat should avoid these
situations. Let's also do it for inserts, even though
it is much less of a problem there.